# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: proxy-server
  version: 0.2.0
  description: |
    Security proxy — jedyny punkt wyjścia agenta. Ruch do LLM (hop A), aplikacji (hop B), MCP i innych agentów
    przechodzi przez proxy.

    Dwie powierzchnie:

    - **Agent API** (`/v1/*`, `/apps/*`) — uwierzytelnienie kluczem agenta `Authorization: Bearer ak_<key_id>_<secret>`.
      Błędy w kształcie `{error: {type, code, message}}`; odpowiedzi decyzji (D10) zawsze niosą `decision_id`
      (także w nagłówku `X-Decision-Id`) i nigdy nie zawierają uzasadnień.
    - **Admin API** (`/api/v1/*`) — control plane dla UI, JWT Supabase. Błędy `{detail}`, błąd walidacji → `400`.

    Każda odpowiedź ma nagłówek `X-Request-Id`.
servers:
  - url: http://localhost:8080
    description: Lokalnie (`uvicorn app.main:app --port 8080`)
...
paths:
...
  /api/v1/agents:
    get:
      tags: [Agents]
      summary: Lista agentów
      operationId: listAgents
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          description: Wiele wartości — powtórzony parametr albo lista po przecinku.
          schema: { type: array, items: { $ref: "#/components/schemas/AgentStatus" } }
          style: form
          explode: true
        - { name: role_id, in: query, schema: { type: string } }
        - name: sort
          in: query
          schema: { type: string, enum: [name, role, status, api_key, last_seen], default: last_seen }
        - $ref: "#/components/parameters/SortDir"
        - $ref: "#/components/parameters/Limit50"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Agents]
      summary: Utwórz agenta (z pierwszym kluczem)
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentCreate" }
      responses:
        "201":
          description: Utworzony; `apiKey` pokazywany tylko raz
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentWithKey" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/agents/{agent_id}:
    parameters: [{ $ref: "#/components/parameters/AgentId" }]
    get:
      tags: [Agents]
      summary: Szczegóły agenta
      operationId: getAgent
      responses:
        "200": { $ref: "#/components/responses/Agent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Agents]
      summary: Edytuj agenta
      operationId: patchAgent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentPatch" }
      responses:
        "200": { $ref: "#/components/responses/Agent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/agents/{agent_id}/revoke:
    parameters: [{ $ref: "#/components/parameters/AgentId" }]
    post:
      tags: [Agents]
      summary: Unieważnij wszystkie klucze agenta
      description: Ustawia status `revoked`. Przywrócenie — przez wygenerowanie nowego klucza.
      operationId: revokeAgent
      responses:
        "200": { $ref: "#/components/responses/Agent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/agents/{agent_id}/keys:
    parameters: [{ $ref: "#/components/parameters/AgentId" }]
    post:
      tags: [Agents]
      summary: Nowy klucz agenta
      description: |
        Klucz pokazywany tylko raz. Stare klucze zostają aktywne do `revoke` (rotacja bez przestoju).
        Agent w statusie `revoked` wraca do `active`.
      operationId: createAgentKey
      responses:
        "201":
          description: Klucz utworzony
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentWithKey" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/agents/{agent_id}/overview:
    parameters: [{ $ref: "#/components/parameters/AgentId" }]
    get:
      tags: [Agents]
      summary: Metryki agenta
      operationId: getAgentOverview
      parameters:
        - $ref: "#/components/parameters/Range"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Tz"
        - $ref: "#/components/parameters/Bucket"
        - $ref: "#/components/parameters/TopLimit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentOverview" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/agents/{agent_id}/posture:
    parameters: [{ $ref: "#/components/parameters/AgentId" }]
    get:
      tags: [Agents]
      summary: Efektywne uprawnienia agenta
      operationId: getAgentPosture
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentPosture" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/agents/{agent_id}/mcp/{server_id}:
    parameters:
      - $ref: "#/components/parameters/AgentId"
      - $ref: "#/components/parameters/ServerId"
    post:
      tags: [Agents]
      summary: Podłącz serwer MCP do agenta (poza MVP)
      operationId: attachAgentMcp
      responses:
        "501": { $ref: "#/components/responses/NotImplemented" }
    delete:
      tags: [Agents]
      summary: Odłącz serwer MCP od agenta (poza MVP)
      operationId: detachAgentMcp
      responses:
        "501": { $ref: "#/components/responses/NotImplemented" }
  /api/v1/agents/{agent_id}/mcp/{server_id}/auth:
    parameters:
      - $ref: "#/components/parameters/AgentId"
      - $ref: "#/components/parameters/ServerId"
    post:
      tags: [Agents]
      summary: Autoryzacja MCP dla agenta (poza MVP)
      operationId: authAgentMcp
      responses:
        "501": { $ref: "#/components/responses/NotImplemented" }
  /api/v1/agents/{agent_id}/rules:
    parameters: [{ $ref: "#/components/parameters/AgentId" }]
    get:
      tags: [Rules]
      summary: Reguły agenta (w kolejności ewaluacji)
      operationId: listRules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Rule" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Rules]
      summary: Dodaj regułę
      description: Bez `position` reguła trafia na koniec listy.
      operationId: createRule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RuleBody" }
      responses:
        "201": { $ref: "#/components/responses/Rule" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/agents/{agent_id}/rules/{rule_id}:
    parameters:
      - $ref: "#/components/parameters/AgentId"
      - { name: rule_id, in: path, required: true, schema: { type: string } }
    put:
      tags: [Rules]
      summary: Zastąp regułę
      description: Bez `position` pozycja zostaje bez zmian.
      operationId: updateRule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RuleBody" }
      responses:
        "200": { $ref: "#/components/responses/Rule" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Rules]
      summary: Usuń regułę
      operationId: deleteRule
      responses:
        "204": { description: Usunięta }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/agents/{agent_id}/rules/meta:
    parameters: [{ $ref: "#/components/parameters/AgentId" }]
    get:
      tags: [Rules]
      summary: Metadane edytora reguł
      description: Toole per aplikacja, pola dostępne w warunkach (z `input_schema`, `args` i faktów) oraz próbki do dry-run.
      operationId: getRulesMeta
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RulesMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/agents/{agent_id}/rules/dry-run:
    parameters: [{ $ref: "#/components/parameters/AgentId" }]
    post:
      tags: [Rules]
      summary: Dry-run reguł + Cedar
      description: |
        Bez `rules` używa zapisanych reguł agenta; z `rules` — szkicu (np. niezapisanych zmian z edytora).
        Gdy `tool` istnieje w katalogu, odpowiedź zawiera też wynik pakietu Cedar (`policy`).
      operationId: dryRunRules
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DryRunBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DryRunResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/agents/{agent_id}/quotas:
    parameters: [{ $ref: "#/components/parameters/AgentId" }]
    get:
      tags: [Quotas]
      summary: Kwoty agenta
      operationId: listQuotas
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Quota" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Quotas]
      summary: Dodaj kwotę
      operationId: createQuota
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuotaCreate" }
      responses:
        "201": { $ref: "#/components/responses/Quota" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/quotas/{quota_id}:
    parameters:
      - { name: quota_id, in: path, required: true, schema: { type: string } }
    patch:
      tags: [Quotas]
      summary: Edytuj kwotę
      operationId: patchQuota
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuotaPatch" }
      responses:
        "200": { $ref: "#/components/responses/Quota" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
security:
  - operatorJwt: []
components:
  securitySchemes:
    agentKey:
      type: http
      scheme: bearer
      bearerFormat: ak_<key_id>_<secret>
      description: Klucz agenta (`python -m app.cli create-key <agent_id>` lub `POST /api/v1/agents/{id}/keys`).
    operatorJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT z Supabase Auth (HS256 z `SUPABASE_JWT_SECRET` albo RS256/ES256/EdDSA przez JWKS).
        Lokalnie przy `ADMIN_AUTH_DISABLED=true` nagłówek jest zbędny; operatora można wskazać przez `X-Dev-Operator: <email>`.
...
  parameters:
...
    AgentId:
      { name: agent_id, in: path, required: true, schema: { type: string, example: purchasing-agent } }
...
    ServerId:
      name: server_id
      in: path
      required: true
      description: Id aplikacji z `proxy.apps`.
      schema: { type: string }
...
    Search:
      { name: search, in: query, schema: { type: string } }
    SortDir:
      name: sort_dir
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
    Limit50:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: Nieprzezroczysty kursor z `nextCursor` (keyset).
      schema: { type: string }
...
    Range:
      name: range
      in: query
      description: Preset okna. Domyślnie `today`; pomiń i podaj `from` + `to` dla zakresu własnego (max 90 dni).
      schema: { type: string, enum: [today, 24h, 7d, 30d] }
    From:
      name: from
      in: query
      description: ISO 8601 ze strefą czasową.
      schema: { type: string, format: date-time }
    To:
      name: to
      in: query
      description: ISO 8601 ze strefą czasową, wyłącznie.
      schema: { type: string, format: date-time }
    Tz:
      name: tz
      in: query
      description: Strefa IANA; domyślnie ze snapshotu ustawień.
      schema: { type: string, example: Europe/Warsaw }
    Bucket:
      name: bucket
      in: query
      description: Domyślnie dobierany do okna.
      schema: { type: string, enum: [5m, 15m, 1h, 1d] }
    TopLimit:
      name: top_limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 5 }
  responses:
...
    BadRequest:
      description: Błędne dane wejściowe
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AdminError" }
    Unauthorized:
      description: Brak / niepoprawny JWT
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AdminError" }
    Forbidden:
      description: Operator wyłączony lub bez uprawnień
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AdminError" }
    NotFound:
      description: Nie znaleziono
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AdminError" }
    Conflict:
      description: Konflikt stanu
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AdminError" }
    NotImplemented:
      description: Poza MVP
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AdminError" }
    Agent:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Agent" }
    Rule:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Rule" }
    Quota:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Quota" }
...
  schemas:
...
    AdminError:
      type: object
      required: [detail]
      properties:
        detail: { type: string, example: Unknown agent }
...
    NextCursor:
      type: [string, "null"]
      description: "`null` = ostatnia strona."

    # ------------------------------------------------------------ agent decisions (D10)
...
    AgentStatus: { type: string, enum: [active, disabled, revoked] }
...
    RoleStatus: { type: string, enum: [draft, active, archived] }
...
    WindowMetrics:
      type: object
      required: [window, calls, callsDeltaPct, denyRatePct, cautionRatePct, rateLimited, decisionSplit, callsOverTime, topTools, pendingApprovals]
      properties:
        window: { $ref: "#/components/schemas/Window" }
        calls: { type: integer }
        callsDeltaPct: { type: [integer, "null"], description: "`null` gdy w poprzednim okresie nie było wywołań." }
        denyRatePct: { type: integer }
        cautionRatePct: { type: integer }
        rateLimited: { type: integer }
        decisionSplit:
          type: array
          items:
            type: object
            required: [decision, count]
            properties:
              decision: { $ref: "#/components/schemas/UiDecision" }
              count: { type: integer }
        callsOverTime:
          type: array
          items:
            type: object
            required: [start, end, count]
            properties:
              start: { type: string, format: date-time }
              end: { type: string, format: date-time }
              count: { type: integer }
        topTools:
          type: array
          items:
            type: object
            required: [tool, count]
            properties:
              tool: { type: string }
              count: { type: integer }
        pendingApprovals: { type: integer }
    QuotaUsage:
      type: object
      required: [id, agentId, label, used, cap, unit]
      properties:
        id: { type: string }
        agentId: { type: string }
        label: { type: string, example: Purchasing Agent · Daily calls }
        used: { type: integer }
        cap: { type: integer }
        unit: { type: string, const: calls }
...
    AgentOverview:
      allOf:
        - $ref: "#/components/schemas/WindowMetrics"
        - type: object
          required: [clear, flagged, clearToday, cautionToday, budget]
          properties:
            clear: { type: integer }
            flagged: { type: integer, description: caution + deny }
            clearToday: { type: integer, description: Alias `clear` (kompatybilność UI). }
            cautionToday: { type: integer, description: Alias `flagged` (kompatybilność UI). }
            budget:
              oneOf:
                - $ref: "#/components/schemas/QuotaUsage"
                - type: "null"
              description: Kwota dzienna, a gdy jej brak — pierwsza aktywna.

    # ------------------------------------------------------------ agents
    Agent:
      type: object
      required: [id, name, roleId, roleName, status, apiKeyHint, mcpServerIds, createdAt, lastSeenAt, mandate, llmModels, rateLimitOverride, limits]
      properties:
        id: { type: string, example: purchasing-agent }
        name: { type: string }
        roleId: { type: [string, "null"] }
        roleName: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/AgentStatus" }
        apiKeyHint: { type: string, example: "ak_1a2b3c4d_••••x9Yz" }
        mcpServerIds: { type: array, items: { type: string }, description: Aplikacje, do których agent ma jakikolwiek grant. }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: [string, "null"], format: date-time }
        mandate: { type: string }
        llmModels: { type: array, items: { type: string }, description: Pusta = dowolny model. }
        rateLimitOverride: { type: [integer, "null"], description: "`limits.requests_per_minute`" }
        limits:
          type: object
          additionalProperties: { type: integer }
          example: { requests_per_minute: 60, max_denies_per_session: 3 }
    AgentWithKey:
      allOf:
        - $ref: "#/components/schemas/Agent"
        - type: object
          required: [apiKey]
          properties:
            apiKey: { type: string, description: Pełny klucz — pokazywany tylko raz., example: ak_1a2b3c4d_secret }
    AgentPage:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Agent" } }
        nextCursor: { $ref: "#/components/schemas/NextCursor" }
    AgentCreate:
      type: object
      required: [id, name, mandate]
      properties:
        id: { type: string, pattern: "^[a-z0-9][a-z0-9-]{1,62}$" }
        name: { type: string, minLength: 1, maxLength: 200 }
        mandate: { type: string, minLength: 1, maxLength: 4000 }
        roleId: { type: [string, "null"] }
        llmModels: { type: array, items: { type: string }, default: [] }
        limits: { type: object, additionalProperties: { type: integer }, default: {} }
    AgentPatch:
      type: object
      additionalProperties: false
      description: Pola pominięte zostają bez zmian. `rateLimitOverride` = `null` usuwa override; `roleId` = `null` odpina rolę.
      properties:
        roleId: { type: [string, "null"] }
        name: { type: [string, "null"], minLength: 1, maxLength: 200 }
        status: { type: [string, "null"], enum: [active, disabled, null] }
        mandate: { type: [string, "null"], minLength: 1, maxLength: 4000 }
        llmModels: { type: [array, "null"], items: { type: string } }
        rateLimitOverride: { type: [integer, "null"], minimum: 1 }
    EffectiveTool:
      type: object
      required: [serverId, serverName, tool, via]
      properties:
        serverId: { type: string }
        serverName: { type: string }
        tool: { type: string, example: marketplace.place_order }
        via: { type: string, enum: [server, tool] }
    AgentPosture:
      type: object
      required: [role, callable, unreachable, attachedWithoutGrants]
      properties:
        role:
          oneOf:
            - $ref: "#/components/schemas/Role"
            - type: "null"
        callable: { type: array, items: { $ref: "#/components/schemas/EffectiveTool" } }
        unreachable: { type: array, items: {}, maxItems: 0 }
        attachedWithoutGrants: { type: array, items: { type: string } }

    # ------------------------------------------------------------ rules
...
    RuleOutcome: { type: string, enum: [allow, deny, needs_ai] }
    RuleBody:
      type: object
      required: [name, tool, then]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        tool: { type: string, minLength: 1, maxLength: 200, description: "Kwalifikowana nazwa toola; dopuszczalne wzorce glob." }
        when:
          $ref: "#/components/schemas/ConditionNode"
          default: { combinator: and, children: [] }
        then: { $ref: "#/components/schemas/RuleOutcome" }
        enabled: { type: boolean, default: true }
        position: { type: [integer, "null"] }
    Rule:
      type: object
      required: [id, name, agentId, tool, when, then, enabled, position, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        agentId: { type: string }
        tool: { type: string }
        when: { $ref: "#/components/schemas/ConditionNode" }
        then: { $ref: "#/components/schemas/RuleOutcome" }
        enabled: { type: boolean }
        position: { type: integer }
        updatedAt: { type: string, format: date-time }
    RulesMeta:
      type: object
      required: [tools, fields, dryRunSamples]
      properties:
        tools:
          type: array
          items:
            type: object
            required: [serverId, serverName, tools]
            properties:
              serverId: { type: string }
              serverName: { type: string }
              tools: { type: array, items: { type: string } }
        fields:
          type: array
          items:
            type: object
            required: [id, label, type, tools]
            properties:
              id: { type: string }
              label: { type: string }
              type: { type: string, enum: [number, enum, text] }
              values: { type: array, items: {} }
              tools: { type: array, items: { type: string }, description: "`[\"*\"]` dla faktów liczonych." }
              computed: { type: boolean }
        dryRunSamples:
          type: array
          items:
            type: object
            required: [id, label, tool, args]
            properties:
              id: { type: string }
              label: { type: string }
              tool: { type: string }
              args: { type: object, additionalProperties: true }
    DryRunBody:
      type: object
      required: [tool]
      properties:
        tool: { type: string }
        args: { type: object, additionalProperties: true, default: {} }
        rules:
          type: [array, "null"]
          items:
            allOf:
              - $ref: "#/components/schemas/RuleBody"
              - type: object
                properties:
                  id: { type: string }
        sessionState: { type: object, additionalProperties: true, default: {} }
        enrichment: { type: object, additionalProperties: true, default: {} }
    DryRunResult:
      type: object
      required: [matchedRuleId, outcome, detail]
      properties:
        matchedRuleId: { type: [string, "null"] }
        outcome: { oneOf: [{ $ref: "#/components/schemas/RuleOutcome" }, { type: "null" }] }
        detail: { type: string }
        policy:
          type: object
          required: [verdict, reasons, facts]
          properties:
            verdict: { type: string, enum: [allow, deny, escalate] }
            reasons: { type: array, items: { $ref: "#/components/schemas/Reason" } }
            facts: { type: object, additionalProperties: true }

    # ------------------------------------------------------------ quotas
    Quota:
      type: object
      required: [id, name, agentId, agentName, window, cap, used, unit, enabled, burst, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        agentId: { type: string }
        agentName: { type: string }
        window: { type: string, enum: [1m, 1h, 1d] }
        cap: { type: integer }
        used: { type: integer }
        unit: { type: string, const: calls }
        enabled: { type: boolean }
        burst: { type: integer }
        updatedAt: { type: string, format: date-time }
    QuotaCreate:
      type: object
      required: [name, window, cap]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        window: { type: string, enum: [1m, 1h, 1d] }
        cap: { type: integer, minimum: 1 }
        burst: { type: integer, minimum: 0, default: 0 }
    QuotaPatch:
      type: object
      properties:
        enabled: { type: [boolean, "null"] }
        name: { type: [string, "null"], minLength: 1, maxLength: 200 }
        cap: { type: [integer, "null"], minimum: 1 }
        burst: { type: [integer, "null"], minimum: 0 }

    # ------------------------------------------------------------ approvals
...
    Role:
      type: object
      required: [id, name, description, status, grants, createdAt, updatedAt, assignedAgents, effectiveTools]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/RoleStatus" }
        grants: { type: array, items: { $ref: "#/components/schemas/RoleGrant" } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        assignedAgents: { type: array, items: { $ref: "#/components/schemas/AgentRef" } }
        effectiveTools: { type: array, items: { $ref: "#/components/schemas/EffectiveTool" } }
...